requirements.txt
Django>=4.2,<5.0
numpy>=1.24
rasterio>=1.3
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
opencv-python>=4.8
Pillow>=10.0
python-dateutil>=2.8
pystac-client
planetary-computer
stackstac
xarray
dask
pyproj
shapely
rasterio
opencv-python
scipy
torch
torchvision